requirements.txt
fastapi>=0.101.0
uvicorn>=0.23.2
sqlmodel>=0.0.8
psycopg2-binary>=2.9.7
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.1.0
email_validator